+++ services/core/shared/core/compat/sdk_v5/accountUtils.js
@@ -17,7 +17,55 @@ const {
 	getAddressFromPublicKey,
 	getBase32AddressFromAddress,
 	getAddressFromBase32Address,
+	getLegacyAddressFromPublicKey,
 } = require('@liskhq/lisk-cryptography');
+
+const coreCache = require('./coreCache');
+
+const accountsIndexSchema = require('./schema/accounts');
+
+const mysqlIndex = require('../../../indexdb/mysql');
+
+const getAccountsIndex = () => mysqlIndex('accounts', accountsIndexSchema);
+
+const parseAddress = address => (typeof address === 'string') ? address.toUpperCase() : '';
+
+const validatePublicKey = publicKey => (typeof publicKey === 'string' && publicKey.match(/^([A-Fa-f0-9]{2}){32}$/g));
+
+const confirmAddress = async address => {
+	if (!address || typeof address !== 'string') return false;
+	const account = await coreCache.getCachedAccountByAddress(address);
+	return (account && parseAddress(account.address) === parseAddress(address));
+};
+
+const confirmPublicKey = async publicKey => {
+	if (!publicKey || typeof publicKey !== 'string') return false;
+	const account = await coreCache.getCachedAccountByPublicKey(publicKey);
+	return (account && account.publicKey === publicKey);
+};
+
+const getIndexedAccountInfo = async params => {
+	const accountsDB = await getAccountsIndex();
+	const [account] = await accountsDB.find(params);
+	return account;
+};
+
+const getAccountsBySearch = async (searchProp, searchString) => {
+	const accountsDB = await getAccountsIndex();
+	const params = {
+		search: {
+			property: searchProp,
+			pattern: searchString,
+		},
+	};
+	const account = await accountsDB.find(params);
+	return account;
+};
+
+const getLegacyFormatAddressFromPublicKey = publicKey => {
+	const legacyAddress = getLegacyAddressFromPublicKey(Buffer.from(publicKey, 'hex'));
+	return legacyAddress;
+};
 
 const getHexAddressFromPublicKey = publicKey => {
 	const binaryAddress = getAddressFromPublicKey(Buffer.from(publicKey, 'hex'));
@@ -34,8 +82,21 @@ const getHexAddressFromBase32 = address => {
 	return binaryAddress;
 };
 
+const getBase32AddressFromPublicKey = publicKey => {
+	const hexAddress = getHexAddressFromPublicKey(publicKey);
+	const base32Address = getBase32AddressFromHex(hexAddress);
+	return base32Address;
+};
+
 module.exports = {
+	validatePublicKey,
+	confirmAddress,
+	confirmPublicKey,
+	getIndexedAccountInfo,
+	getAccountsBySearch,
+	getLegacyAddressFromPublicKey: getLegacyFormatAddressFromPublicKey,
 	getHexAddressFromPublicKey,
 	getBase32AddressFromHex,
 	getHexAddressFromBase32,
+	getBase32AddressFromPublicKey,
 };
